implementations/sme_capitalAllocation/pyproject.toml
[project]
name = "sme-capital-eval"
version = "0.1.0"
description = "NXTFrontier SME Capital Allocation — Agentic Evaluation System"
requires-python = ">=3.12,<3.13"
dependencies = [
    "crewai>=1.6.1",
    "google-genai>=1.67.0",
    "json-repair>=0.25.0",
    "pandas>=2.3.2",
    "numpy>=1.26,<2.0",
    "requests>=2.33.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.2.2",
    "datasets>=4.0.0",
    "langfuse>=4",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
constraint-dependencies = [
    "cryptography>=46.0.6",
    "requests>=2.33.0",
]
